+++ c/debug.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
 set -e
 
-. ../scripts/make_tl.sh
-make_tl DEBUG=1 "${@}"
+../template/make_tl.sh DEBUG=1 "${@}"